+++ components/form/FormWizard.js
@@ -1,11 +1,11 @@
 import React, { useState } from 'react';
 import { Form, Formik } from 'formik';
-import store from 'store2';
+import { animateScroll as Scroll } from 'react-scroll';
 
 const FormWizard = ({ children, initialValues, onSubmit }) => {
   const [stepNumber, setStepNumber] = useState(0);
   const [snapshot, setSnapshot] = useState(initialValues);
-  const [confirmationMessage, setConfirmationMessage] = useState();
+  const [submissionMessage, setSubmissionMessage] = useState(null);
 
   const steps = React.Children.toArray(children);
   const step = steps[stepNumber];
@@ -13,23 +13,33 @@ const FormWizard = ({ children, initialValues, onSubmit }) => {
   const isLastStep = parseInt(stepNumber) === parseInt(totalSteps) - 1;
 
   const next = (values) => {
-    store.set('demo_form', values);
     setSnapshot(values);
     setStepNumber(Math.min(parseInt(stepNumber) + 1, totalSteps - 1));
+    Scroll.scrollToTop();
   };
 
   const previous = (values) => {
     setSnapshot(values);
     setStepNumber(Math.max(parseInt(stepNumber) - 1, 0));
+    Scroll.scrollToTop();
   };
 
   const handleSubmit = async (values, bag) => {
     if (isLastStep) {
-      // save will go here
-      return onSubmit(values, bag);
+      const response = await fetch('/api/save-form', {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify(values),
+      });
+      const data = await response.json();
+      setSubmissionMessage(data.message);
+      response.status === 401 ? null : (bag.resetForm(), setStepNumber(0));
     } else {
       bag.setTouched({});
       next(values);
+      setSubmissionMessage(null);
     }
   };
 
@@ -66,11 +76,8 @@ const FormWizard = ({ children, initialValues, onSubmit }) => {
                     </button>
                   </div>
                 </div>
-                {confirmationMessage === false ? (
-                  <div className='notification is-danger mb-5'>
-                    <button className='delete'></button>
-                    Submission error, please try again
-                  </div>
+                {submissionMessage !== null ? (
+                  <div className='notification mb-5'>{submissionMessage}</div>
                 ) : null}
               </div>
             </section>