+++ pages/Profile/profile.js
@@ -1,11 +1,13 @@
-import { useEffect } from 'react';
+import { useEffect, useState } from 'react';
 import Router from 'next/router';
 import Head from 'next/head';
 import MetaDefaults from '../../components/MetaDefaults';
 import { useAuth } from '../../utils/auth/useAuth';
+import Card from '../../components/Card';
 
 const Profile = () => {
   const { user, isLoggedIn } = useAuth();
+  const [forms, setForms] = useState([]);
 
   // protect profile from unauthorized users
   useEffect(() => {
@@ -13,6 +15,20 @@ const Profile = () => {
       Router.push('/');
     }
   }, [isLoggedIn]);
+
+  useEffect(() => {
+    (async () => {
+      const response = await fetch('/api/get-forms', {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify(user ? user.email : null),
+      });
+      const data = await response.json();
+      setForms(data);
+    })();
+  }, []);
 
   if (isLoggedIn) {
     return (
@@ -32,18 +48,29 @@ const Profile = () => {
         <MetaDefaults />
         <section className='section content'>
           <div className='container'>
-            <h2 className='title is-2 has-text-centered'>
-              {`Hello ${user.email}!`}
-            </h2>
-            <p className='has-text-centered'>
-              Glad you made it! This is a sample profile page, see your
-              personalized welcome up above?
-            </p>
-            <p className='has-text-centered'>
-              If you submitted a form and logged in with the same email address
-              you input within the form, you can see that data here. If you
-              haven't used the form yet, go ahead and try it out!
-            </p>
+            <div className='columns'>
+              <div className='column is-12'>
+                <h2 className='title is-2 has-text-centered'>
+                  {`Hello ${user.email}!`}
+                </h2>
+                <p className='has-text-centered'>
+                  Glad you made it! This is a sample profile page, see your
+                  personalized welcome up above?
+                </p>
+                <p className='has-text-centered'>
+                  If you submitted a form and logged in with the same email
+                  address you input within the form, you can see that data here.
+                  If you haven't used the form yet, go ahead and try it out!
+                </p>
+              </div>
+            </div>
+            <div className='columns is-centered is-multiline'>
+              {/* <div className='column is-3'> */}
+              {forms.map((item, i) => {
+                return <Card key={i} {...item} />;
+              })}
+              {/* </div> */}
+            </div>
           </div>
         </section>
       </div>