+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+PR Add form data functionality - 10/14/2021
+
+- Add api route and functionality to save form
+- Add api route and functionality to fetch forms
+- Add form display to user profile
+- Fix misc bugs
+
 PR Add user profile, logout - 10/13/2021
 
 - Add user profile page